Refactor : 여정 조회부분 장소 댓글 MVVM 모델 적용
장소 댓글 부분에 MVVM 모델을 적용하여 가독성을 높임
